Initial commit
At this point the main class SurveyWidget holds a master detail style
relationship between CategoryList.java and other Detailed interfaced
Composite subclasses such as TextCategoryComposite and
BasicPainCategoryComposite. In this next build I will be trying to
modify the JSON objects with a mutable PainCategory one, so that the
Detail interface can interact with the Master class.